New ngram class is included now. affine.hxx:Strings are appended to now. Static casts used. affine_test.cpp: Make sure that I cast to an int so that it's printed right. hill_cipher.hxx: Removed some old commented out code. math.hxx:Fixed up some formatting. ngram.hxx: New Ngram class. ngram_test.cpp: New test driver code. CMakeLists.txt: Make "debug" code just be basic optimizations. Prod is O2 optimizations.
Lots and lots of fixes in the code now.

Ngram class is now live and seems to produce results exactly like I want. Commented out version is a "map" or a BST which isn't what I care about when in memory. The hashtable is exactly what I want. When it comes to printing it out I'll have to figure out something. But the ngram values aren't normally ever going to be printed out in any of the normal use-cases.